+++ CHANGELOG.md
@@ -8,7 +8,7 @@ Add support for hierarchical range filters (#154)
 
 
 #### Changed
-
+- Remove active checking for blob update (#159)
 
 #### Fixed
 - corrupted blob now should be saved in case of 'unexpected eof' error